% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_convert.r
\name{xmlconvert}
\alias{xmlconvert}
\title{Package 'xmlconvert'}
\description{
Tools for converting XML documents to dataframes and vice versa

Functions available:
\itemize{
\item \code{\link{xml_to_df}()}: Converts an XML document to an R dataframe
\item \code{\link{df_to_xml}()}: Converts a dataframe to an XML document
}
}
